Fixes circular dependency with queries
