+++ components/BlockModalMenu/fragments/blockModalMenuChannel.js
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag'
+
+export default gql`
+  fragment BlockModalMenuChannel on Channel {
+    id
+    title
+    visibility
+  }
+`
