+++ components/BlockModalMenu/mutations/deleteConnection.js
@@ -1,0 +1,10 @@
+import gql from 'graphql-tag'
+
+export default gql`
+  mutation deleteConnectionMutation($id: ID!) {
+    delete_connection(input: { id: $id }) {
+      clientMutationId
+      status
+    }
+  }
+`
